requirements.txt
pysaucenao ~= 0.2.1
discord.py ~= 1.3.3
aiohttp
